[BLD] Update deploy automation to deploy control and data planes separately (#4917)
## Description of changes

Updates the places in our GitHub Actions where we dispatch the
`deploy.yaml` workflow in our internal repository to match the updates
to its inputs. The deploy workflow was updated to take a single `plane`
string value instead of a comma-separated `planes` input. The update on
the other side is intended to affect the deployments to production to
further reduce the likelihood that one of the planes is inadvertently
deployed.

Note: Rather than write the throw-away code on the other side to support
both `plane` and `planes` inputs, I'm going to land this change shortly
after landing the change in hosted, but there may be older versions of
the `trigger-deploy.yaml` workflow that fail if they dispatch with the
old `planes` input.

## Test plan

Once this change lands, it will trigger a deploy and we will see if it
works.