[TST] Add tests for CheckCollections, make test collection setup more flexible (#4912)
## Summary
This adds a test for the `CheckCollections` method. This would have
caught the issue fixed in #4899.

Along the way, I discovered how these tests need some love, so I started
down this path by introducing a new `daotest` package that provides
testing helpers for creating `Collection` values that are easier to
configure. Instead of having to potentially debug all the tests that use
`dao.CreateTestCollection`, I defined a "shim" that can be used as a
stop-gap so that I can change the signature of this
`dao.CreateTestCollection` without potentially derailing the original
goal which was just to add some missing test coverage.

- Improvements & Bug fixes
  - Added test coverage for `CheckCollections` service endpoint
- New functionality
- Added `daotest` package and `NewTestCollection` helper with
builder/options methods for more configurability

## Test plan

All the tests are passing.

## Documentation Changes

N/A