[BUG][wal3]  GC gets wedged
When a log contention error is encountered in GC, it will proceeed to
delete without installing the manifest.  Further, if an existing GC file
is found after this case, the GC will be wedged trying to make garbage
from a snapshot that doesn't exist.

The fix is to retry applying the manifest until there is no garbage
work.  And then try loading garbage before generating garbage.