[ENH]: improve ListCollectionsToGc with filter for min alive versions (#4971)
## Description of changes

There is a large number of collections in staging right now which are
alive and have < 3 versions. They are blocking eligible collections from
being GC'ed; namely deleted collections with < 2 versions are currently
blocked because the former set appears first and there is a `LIMIT
1000`.

This updates the query on SysDb so that only collections actually
eligible for GC will be returned.

## Test plan

_How are these changes tested?_

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_

n/a