[BLD] Optimize which tests we run on Windows (#4955)
## Description of changes

We want to optimize our usage of Windows runners since they cost so
much. The cost is due in large part to the amount of time our individual
jobs take to run against Windows, but also the sheer number of jobs we
run on Windows when we merge to main or push a tag. Details are below:

- Improvements & Bug fixes
- Only run the Windows tests against the latest version of Python --
this greatly reduces the number of jobs we run on Windows for a single
merge to main (reduces the total number of jobs by ~3x). The logic being
that when we see failures for Python tests, we see them across all
versions of Python and not just a single version _AND_ we are running
the tests against the full matrix on Linux anyways.
- Optimizes the logic for when we run the smoke tests -- we were running
them on merges to main and tags when we shouldn't have been (it's
redundant).

## Test plan

This doesn't change the tests, it just changes how many tests we are
running. I've also double/triple tested/checked the logic for the smoke
tests and I am pretty sure I got it right this time.


## Documentation Changes

N/A