[CHORE]: use a match for defining default quotas (#4941)
## Description of changes

This PR introduces a trait `DefaultQuota` that is implemented for
`UsageType` as a way to unfailingly get a default quota. Since default
quotas are known at compilation, we should have an unfailing way to get
them for the UsageType enum. We still need the `DEFAULT_QUOTAS` hashmap
because this mimics the data structure in which tenant overrides are
stored.

## Test plan

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A