[ENH] Pipeline compactions for different collections (#4963)
## Description of changes

Before this change, compactions would be done in batches of collections.
This would mean each batch would take as long as the heaviest compaction
job in its batch, making it very easy for a heavy compaction to stall
out compactions in later batches. This change reworks the compaction
jobs such that they can run asynchronously with respect to each other.

This change's contributions are as follows:
- Every time the scheduler fetches a bunch of collection ids to start
compactions on, the CompactionManager creates a future for each of these
that gets sent to a helper thread who awaits on these. Each time this
helper thread finds one of these futures to be completed, it checks to
see if the scheduler and CompactionManager have produced any more tasks
in the meantime and starts them.
- A deduplication map of in-progress compaction jobs is maintained to
make sure the same job doesn't execute in multiple concurrent instances.
The helper thread described above sends job completion messages to the
CompactionManager that can be handled to remove entries from this map.
Each entry has an expiry time in the map of at one hour for now in case
the helper thread dies for any reason. This is supposed to be a failure
mode, not an expected case but the expiration mechanism provides a
helpful fallback in such a scenario.

- Improvements & Bug fixes
- A local workload with 100 collections and 1000 records writes per
collection takes ~40s to finish its writes and compactions. Before this
change the presence of another concurrent workload writing 10k records
to one collection caused the 40s to shoot up to ~60s. After this change
said workload's timing does not go above ~42s in any statistically
significant manner in the presence of the concurrent heavy workload.
- New functionality
  - N/A

## Test plan
Perf testing described above.

- [v] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_

---------

Co-authored-by: Tanuj Nayak <tnayak@Tanujs-MacBook-Pro.local>
Co-authored-by: Tanuj Nayak <tnayak@mac.lan>