[ENH]: return error on /add if embeddings are not provided (#5033)
## Description of changes

Our official clients prevent this but the server does not have a check.
A better solution is making `AddCollectionRecordsPayload.embddings`
non-optional, but this works as a minimal patch for now.

## Test plan

_How are these changes tested?_

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_

n/a