[BLD] Fix windows runner label (#5052)
## Description of changes

In #5027 I accidentally wrote the Windows runner label as
"8core-64gb-windows-latest" when it should be
"8core-32gb-windows-latest" (32gb vs 64gb). I have since created a
runner that matches the wrong label, but this was causing our windows
builds to back up.

## Test plan

We will see the release job run when this is merged.