[CHORE] Add pprof server to query service (#5072)
## Description of changes

Similar to what we did in #4845, this adds the jemalloc pprof server to
the query-service server so that we can do memory profiling for the
service using Polar Signals.

Note: I also updated a place in the garbage collector's implementation
where the port was hard-coded to 6060.

- New functionality
  - pprof server started when config given

## Test plan

Build should succeed and updating `chroma_config.yaml` with local tilt
environment to test that pprof endpoint is up.

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust