[CHORE]: hide private service names in otel/tracing init (#5075)
## Description of changes

Currently, the OSS code is responsible for initializing tracing/otel for
all crates. A global filter is applied in which for all crates that
initialize tracing, every filter is applied for that crate. This PR
slightly refactors this pattern to define a set of "global default crate
filters" and allow initializers to pass in additional crate filters (and
even overrides) at their call sites.

## Test plan

This PR is the first in a stack with
https://github.com/chroma-core/hosted-chroma/pull/3345. Together, these
refactor the tracing and otel initialization to move private service
names into our private code. Forward/backward compatibility would
require a dance in which we doubly-initialize filters for the same
crates, which feels like more complexity than it's worth given the scope
of these changes. Therefore, these two PRs should be rolled out
together.

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

N/A