[CHORE]: update tracing filters to use underscores instead of hyphens (#5111)
## Description of changes

The tracing issues we've been chasing boil down to a _fun_ little quirk
of Rust compiler, in which hyphens in crate names are converted to
underscores automatically to make the crate names valid Rust
identifiers.

## Test plan

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan

N/A

## Observability plan

Monitor staging to make sure missing traces start to reappear.

## Documentation Changes

N/A