[ENH]: Leader election for SysDB (#5104)
## Description of changes

This change introduces leader election in the SysDB service in
preparation to scale out this layer. After this change, only the leader
monitors and updates the member list. Not that the leader election in
question just reuses the Kubernetes leader election API so it is not
strongly consistent and can lead to split-brain scenarios. These
split-brain scenarios are not harmful here.

Since this change reuses the leader election code that's already in the
log service, that code has been moved out of the log service directory.

- Improvements & Bug fixes
  - Gated member list service operations by a leader election.
- New functionality
  - Described above. This should enable the SysDB service to scale out.

## Test plan

Observing log statements on local tilt.

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan

N/A. Before scaling out the SysDB layer, make sure this change is in the
SysDB binary.

## Observability plan

N/A

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_