[BUG] ci: Use correct step dependencies for Slack notifications (#5149)
## Description of changes

Between the time I originally created this PR and when I revisited it
and eventually merged it, the steps upon which the
`notify-slack-on-failure` step depends on needed to be updated, which I
didn't do, so the workflow is technically invalid. This fixes it by
using the correct step dependencies.

- Improvements & Bug fixes
- Uses the correct step dependencies in the `notify-slack-on-failure`
step

## Test plan

I validated against the GitHub Action schema linter. It was showing the
bug right away.

## Observability plan

We should see this run after the change merges.

## Documentation Changes

N/A