[BUG]  Snapshot race condition leads to corrupt manifest.
There's a race condition in which the path from the root to the earliest
retained fragment will be updated in the manifest to put the fragment in
a deeper, path, and thus not be the same set of snapshots to cull.

The fix is to fail to apply garbage collection if there is a missing
snapshot and retry the entire GC operation.

Additionally, it was possible for GC to get in a state where the garbage
was taken, but not applied.  Add a channel for communicating the
success back to the waiter.

Also includes:
- Proptest for log GC under contention.
- Unit test to test snapshot rollup after GC.
- Add logic to remove garbage file when destroying log.
- Update the branching factor of manifest in proptest to small values to
  shrink the number of logs needed for proptest.
- Update scrub logic to make use of the oldest_timestamp() calculation.
- Removed an assert that panics on valid scenarios.
- Removed a broken unit test.