[ENH]  Do not trace failed preconditions when flushing compaction.
The soft-deleted compaction will return an error that's not
distinguishable from other errors.  It will percolate to the dispatcher
and be logged like every other error.

In #5064, we added a should_trace_error function on the chroma Error
trait, which will silence the error at the dispatcher.  This PR applies
that annotation to the failed precondition.