[ENH] Move fetch to its own task for cancellation (#5158)
## Description of changes

_Summarize the changes made by this PR._

- Improvements & Bug fixes
- This moves the actual network request into its own tokio task so that
it if a parent task is cancelled the request can drive forward and still
make progress for other callers.
- Even if all callers are dropped, we still allow the request to make
progress
- The spawned task is responsible for removing the task from the map,
only under a panic will it not.
- If any waiters are cancelled the send() error to them is ignored and
currently traced as an error as this should only happen under timeout
situations. We can downgrade to a warn in the future
  - 
- New functionality
  - ...

## Test plan

_How are these changes tested?_
Existing tests ensure the subsystem works, this is a non-functional
change
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan
None

## Observability plan
Added error traces for timeout/cancel conditions, can be downgraded to a
warning
 
## Documentation Changes
None

---------

Co-authored-by: propel-code-bot[bot] <203372662+propel-code-bot[bot]@users.noreply.github.com>