[BUG] Use InternalUpdateConfiguration in Rust, correctly update configuration in go (#5069)
## Description of changes

Previously, the go code did not update the configuration str for
collections because it was not passed as part of
`generateCollectionUpdatesWithoutID`

This was fixed, and tests were added. It also cleans up the existing go
types, and updates the rust types to use a similar
InternalCollectionConfiguration for updates.



## Test plan

_How are these changes tested?_

Tests were added to validate that both the modify worked, and on future
fetches maintains the values. Db tests were also added to the dao

- [x ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_