[ENH]  Add a hack to allow us to move past some broken collections on staging.
This introduces a new configuration for the purposes of skipping the
num-versions-to-keep rule in GC for the wal3 log only.  When the new
enable_dangerous_option_to_ignore_min_versions_for_wal3 is set, it will
try to do the right thing, and if it fails with the symptoms of the bug,
it will then try again using just the compaction offset.