[ENH]  Add a hack to allow us to move past some broken collections on staging. (#5174)
## Description of changes

This introduces a new configuration for the purposes of skipping the
num-versions-to-keep rule in GC for the wal3 log only.  When the new
enable_dangerous_option_to_ignore_min_versions_for_wal3 is set, it will
try to do the right thing, and if it fails with the symptoms of the bug,
it will then try again using just the compaction offset.

## Test plan

Send through CI, enable on staging, then stop.

## Migration plan

This is a migration of sorts.

## Observability plan

I will watch staging where we currently see errors and will not see them
in the future.

## Documentation Changes

N/A