[ENH]  And endpoint and tool to rollback collection log offset after DR.
After Disaster Recovery (DR) we may be in an inconsistent state where
the log is ahead of the sysdb.  This is not a problem for safety as the
log will remain durable; the compaction will fail with the log being
behind.

This PR mitigates the problem by introducing an endpoint that can roll
back the cursor.  The current endpoint will silently succeed if the
endpoint is rolled back.  The new endpoint will rollback unconditionally
to match the sysdb offset.