[ENH]  When the log offset is behind sysdb, repair. (#5179)
## Description of changes

The short of it is that this error happens when the log service and the
sysdb state disagree about the latest compaction for data.  Because the
sysdb is the source of truth, we write to the sysdb and then write to
the log service.  This is not atomic.  Breaking the appearance of
atomicity is the first and longstanding way that this problem can
happen.

The other is a new failure mode introduced with the rust log service.
Because it has more than one replica and there's only loose binding
between replicas and logs, it becomes possible for one log to service a
request only to have another immediately service the same log.  This
leads to both having the dirty marker for the log, but only one will
receive the `update_collection_log_offset` call that clears the marker.

This PR addresses both failure modes by calling
update_collection_log_offset on every rust log service node when the
problem happens.

## Test plan

CI and staging

## Migration plan

N/A

## Observability plan

We should see fewer instances of the problem.  They show up in logs.

## Documentation Changes

N/A

---------

Co-authored-by: propel-code-bot[bot] <203372662+propel-code-bot[bot]@users.noreply.github.com>