[ENH]  Beef up recent test against flakes. (#5184)
## Description of changes

The recent change to the rust log service and compactor added the
ability to repair automatically when the log is behind sysdb.  The test
flaked on landing, so this patch is to make it not flake.

Check that it rolls back the offset and give more time to repair.

## Test plan

Flaked on landing.  Have run ten in a row locally without flake.

CI.

## Migration plan

N/A

## Observability plan

N/A

## Documentation Changes

N/A