[ENH]: Adaptive search nprobe based on collection sizes (#5185)
## Description of changes

_Summarize the changes made by this PR._

- Improvements & Bug fixes
- Introduces a config adaptive_search_nprobe that is set to true by
default
- If adaptive_search_nprobe is false then it uses the search_nprobe
present in collection config
- If adaptive_search_nprobe is true then it uses a rule based logic for
determining nprobe. The exact sequence of if-else is left to reading of
the code
- It always ensures that number of points queried is 20x of user
supplied k
- Cleans up the config to not be optional anymore since readers also use
it now
- New functionality
  - ...

## Test plan

_How are these changes tested?_
We are doing experiments on various datasets to test the minimum
possible search_nprobe as a function of the dataset size that gives an
acceptable recall. This change is informed by those experiments
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan
None

## Observability plan
Doing experiments that measure recall to ensure that it does not
degrade.
_What is the plan to instrument and monitor this change?_

## Documentation Changes
None