[ENH]  The rust log service uses connect_lazy; change to connect strict (#5186)
## Description of changes

I hypothesize that the connect_lazy code path (only take for log service
clients) is to blame for our hangs when there's churn in the rust log
service on staging.  Change to connect() rather than connect_lazy() so
the whole grpc handshake gets done before trying to use the connection.

## Test plan

CI

## Migration plan

N/A

## Observability plan

I plan to deploy to staging and then beat it up until the problem occurs
or the problem is doesn't manifest and is therefore likely to be gone.

## Documentation Changes

N/A