[BUG]  Do not ever fail phase2 gc with a contention error. (#5197)
## Description of changes

We were calling heartbeat at the top of the phase2.  This can only
introduce problems now that it lives within the rls rather than on the
garbage collector.

This was caught by proptest failure and fixed by changing the
obvious/committing the seed.

## Test plan

Local + CI

## Migration plan

N/A

## Observability plan

We should see fewer of this error on the garbage collector in staging.

## Documentation Changes

N/A