[BUG] re-enter write metering context on delete (#5195)
## Description of changes

Delete operations should result in two metering events being flushed,
both a read and a write. Traces indicate that write events (the latter
in the sequence) are failing to be flushed. Debugging reveals that the
active metering context on the current thread at the time of flush for
write events is in fact a read context. This happens because the parent
`retryable_delete` future's metering context is a read context, since we
need to specify *a* metering context container when creating a future we
wish to meter.

The result is that when we call `self.log_client.push_logs(/* ...
*/).await`, and this future gets executed on the same thread as our
current future, it inherits the read context from its parent future and
overwrites the write context we manually entered. This results in us
attempting to access a write context when the active context is actually
a read.

The solution is to wrap the push logs call in the write context and
manually re-enter the write context after the push logs call completes.

## Test plan

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan

N/A

## Observability plan

Monitor Argo logs and Honeycomb to make sure downcast errors are gone.

## Documentation Changes

N/A