[ENH]  Handle when the garbage doesn't overlap the manifest.
If the same garbage file gets picked up twice, the second time will
result in a 00000 != XXXXX setsum mismatch.  This is benign, but
visible.  The core issue is that no snapshots get collected, no
fragments get collected, and therefore a 00000 setsum is issued.

To fix it, allow a garbage file with no snapshots in the manifest, whose
fragments fall below the initial seq no of the manifest, to simply pass
as if it was GC'ed.  This was the original intent, it just wasn't
tested.  So add a test too.