[ENH]  Wire up the should_trace_error for soft-deleted collections. (#5228)
## Description of changes

Turns out the previous attempt to fix this did not do so.  I needed to
propagate the should_trace_error call from containing type to contained
type.  This PR does that and adds some tests to assure it will be the
last word.  Let's hope it is.

## Test plan

Local + CI.

It adds tests, so short of wiring up tracing in CI and looking for an
error on a compaction of a soft-deleted collection, this is the best
we'll do.

## Migration plan

N/A

## Observability plan

It currently pages.  It shouldn't.

## Documentation Changes

N/A