[ENH]: Make number of concurrent block flushes configurable (#5252)
## Description of changes

_Summarize the changes made by this PR._

- Improvements & Bug fixes
- PR https://github.com/chroma-core/chroma/pull/5231 changed the
behavior of flushes to flush n blocks at a time instead of flushing all
of them at once. This PR makes this value of n configurable. A new
config field `num_concurrent_block_flushes` is introduced in the block
manager config that the flusher uses to implement this behavior.
- This PR also turns off adaptive_nsearch in CI since it flakes in CI.
It is already off in staging and prod
- New functionality
  - ...

## Test plan

_How are these changes tested?_
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan
None

## Observability plan
Local config tests pass

## Documentation Changes
None