[ENH]: Fix test_add flake (#5272)
## Description of changes

_Summarize the changes made by this PR._

- Improvements & Bug fixes
- test_add.py was flaking because of stack overflow in the spann
compaction code. On deeper investigation it was found that the number of
points added were a few hundred and the spann params were very high for
this data size. Thus, it cascaded into a constant append, split,
reassign, split and so on causing the stack depth to be very high.
- Reduce the params write_nprobe and reassign_neighbor_count to 4 for
test and it succeeds.
- New functionality
  - ...

## Test plan

_How are these changes tested?_
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan
None

## Observability plan
CI should be green

## Documentation Changes
None